made seeder.js readable by separating logic into util functions, created config.js in /server, added fetEnvVariable function and adjusted .env.example
