+++ server/.env.example
@@ -1,3 +1,5 @@
+# DONT FORGET TO UPDATE config.js WHEN ADDING NEW ENV VARIABLES
+
 # This is the port our server will be running on
 PORT=3000
 
@@ -10,9 +12,18 @@ SALT_ROUNDS=12
 # expiration time for email validation tokens in seconds
 EMAIL_VALIDATION_EXPIRATION=600
 
+# Seeder options
+DROP_DB=true# if true, the database will be dropped before seeding
+NUM_USERS=50
+MAX_NUM_POSTS_PER_USER=10
+AVG_NUM_LIKES=5# average 5 likes per post
+AVG_NUM_COMMENTS=3# average of 3 comments per post
+AVG_NUM_FOLLOWS=5# average of 5 follows per user
+
 # When running cypress tests, make sure to change the database. Our cypress will only seed/clean databases that are called cypressDatabase to protect our normal data
 # MONGODB_URL=mongodb+srv://<name>:<password>@<cluster>/cypressDatabase?retryWrites=true&w=majority
 
 # THIS IS THE EMAIL AND PASSWORD FOR THE EMAIL ACCOUNT WE USE TO SEND EMAILS FROM OUR SERVER
-OUR_EMAIL=example@gmail.com
-OUR_APP_PASSWORD=examplepassword
+EMAIL_PROVIDER=gmail
+EMAIL=example@gmail.com
+EMAIL_PASSWORD=examplepassword
