+++ server/src/db/connectDB.js
@@ -1,5 +1,6 @@
 import mongoose from "mongoose";
+import config from "../config.js";
 
-const connectDB = () => mongoose.connect(process.env.MONGODB_URL);
+const connectDB = () => mongoose.connect(config.MONGODB_URL);
 
 export default connectDB;
